Implemented "using" statement for all IDisposable Sqlite Commmands
All sqlite commands were converted to the C# "using" keyword in order to
provide conciseness and proper usage of IDisposable without using a try
catch finally statement (which most methods did NOT implement).
Furthermore, disposed of the sqlite connection in the Dispose(true)
method instead of just closing and setting the variable to null, to
prevent memory leaks.